Use release actual state instead of CR state
