+++ Gopkg.toml
@@ -53,7 +53,6 @@ required = [
   name = "github.com/giantswarm/apprclient"
 
 [[constraint]]
-  branch = "master"
   name = "github.com/giantswarm/e2e-harness"
 
 [[constraint]]
@@ -61,7 +60,7 @@ required = [
   name = "github.com/giantswarm/e2etemplates"
 
 [[constraint]]
-  branch = "master"
+  branch = "list-releases"
   name = "github.com/giantswarm/helmclient"
 
 [[constraint]]
