+++ service/collector/collector.go
@@ -3,7 +3,6 @@ package collector
 import (
 	"sync"
 
-	"github.com/giantswarm/apiextensions/pkg/clientset/versioned"
 	"github.com/giantswarm/helmclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
@@ -19,21 +18,16 @@ const (
 )
 
 type Config struct {
-	G8sClient  versioned.Interface
 	HelmClient *helmclient.Client
 	Logger     micrologger.Logger
 }
 
 type Collector struct {
-	g8sClient  versioned.Interface
 	helmClient *helmclient.Client
 	logger     micrologger.Logger
 }
 
 func New(config Config) (*Collector, error) {
-	if config.G8sClient == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.G8sClient must not be empty", config)
-	}
 	if config.HelmClient == nil {
 		return nil, microerror.Maskf(invalidConfigError, "%T.HelmClient must not be empty", config)
 	}
@@ -42,7 +36,6 @@ func New(config Config) (*Collector, error) {
 	}
 
 	c := &Collector{
-		g8sClient:  config.G8sClient,
 		helmClient: config.HelmClient,
 		logger:     config.Logger,
 	}
@@ -51,7 +44,7 @@ func New(config Config) (*Collector, error) {
 }
 
 func (c *Collector) Describe(ch chan<- *prometheus.Desc) {
-	ch <- chartConfigDesc
+	ch <- releaseDesc
 	ch <- tillerReachableDesc
 }
